Move the new menu into the 'Menu/' directory
